moving everything to ffmpeg that needs to be
