+++ griptape_nodes_library/video/split_video.py
@@ -16,13 +16,16 @@ from griptape_nodes.files.file import File
 from griptape_nodes.retained_mode.griptape_nodes import GriptapeNodes, logger
 from griptape_nodes.traits.options import Options
 
-from griptape_nodes_library.utils.ffmpeg_utils import get_ffmpeg_paths, run_ffmpeg_cmd
+from griptape_nodes_library.utils.ffmpeg_utils import (
+    build_video_segment_cmd,
+    detect_video_properties,
+    get_ffmpeg_paths,
+    run_ffmpeg_cmd,
+)
 from griptape_nodes_library.utils.video_utils import (
     MIN_VIDEO_FILE_SIZE,
     VIDEO_DURATION_BUFFER,
-    build_video_segment_cmd,
     detect_video_format,
-    detect_video_properties,
     sanitize_filename,
     smpte_to_seconds,
     to_video_artifact,
